requirements.txt
python-dotenv
llama-index
Flask
streamlit
PyPDF2
python-docx
openai